+++ Server/typressServer/typressServer.csproj
@@ -72,10 +72,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PrintLogger.cs" />
+    <Compile Include="ViewHandler.cs" />
+    <Compile Include="PrintHandler.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SocketHandler.cs" />
+    <Compile Include="ThreadHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
